+++ client/src/stores/user.js
@@ -4,7 +4,8 @@ import { getMailUnreadCount } from '@/api/mailbox'
 import { getDetails } from '@/api/user'
 import serverData from '@/helper/server-data'
 
-const rateLimitInterval = 300000 // 5 minutes in milliseconds
+const mailUnreadCountRateLimitInterval = 300000 // 5 minutes in milliseconds
+const userDetailsRateLimitInterval = 60000 // 1 minute in milliseconds
 
 export const store = Vue.observable({
   mailUnreadCount: 0,
@@ -101,17 +102,23 @@ export const getters = {
 
 export const mutations = {
   async fetchDetails () {
+    const cacheRequestName = 'userDetails'
     try {
-      store.details = await getDetails()
+      if (await getCacheInterval(cacheRequestName, userDetailsRateLimitInterval)) {
+        store.details = await getDetails()
+
+        await setCache(cacheRequestName, store.details)
+      } else {
+        store.details = await getCache(cacheRequestName)
+      }
     } catch (e) {
-      store.details = null
+      console.error('Error fetching user details:', e)
     }
   },
-
   async fetchMailUnreadCount () {
     const cacheRequestName = 'mailUnreadCount'
     try {
-      if (await getCacheInterval(cacheRequestName, rateLimitInterval)) {
+      if (await getCacheInterval(cacheRequestName, mailUnreadCountRateLimitInterval)) {
         const unreadCount = await getMailUnreadCount()
         store.mailUnreadCount = unreadCount